General Refactor.

+++ trunk/GA/src/edu/itba/ia/tp1/engine/population/selection/UniversalImpl.java
@@ -19,7 +19,7 @@ public class UniversalImpl implements I_SelectionAlgorithm {
 	 * @see edu.itba.ia.tp1.motor.population.IPopulationAlgorithm#execute()
 	 */
 	@Override
-	public Population execute(Population population, I_Aptitude aptitude,Integer nIndividuals) {
+	public Population execute(Population population, I_Aptitude aptitude, Long nIndividuals) {
 		return null;
 	}
